+++ index.html
@@ -39,7 +39,7 @@
   </div>
 </div>
 
-<script type="module" src="app.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
